feat!: enhance HTML attribute parsing and representation
This commit refactors the HTML attribute parsing and representation in the Node struct. It introduces a new Attributes struct and an AttributeValues enum to better handle different attribute data types. It also updates the corresponding code in the parser and to_md modules, as well as the relevant tests. This change provides a more robust and flexible way to handle HTML attributes, improving the overall quality and maintainability of the code.

Release-As: 0.8.0